+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Auth datacube `earthdaily.earthdatastore.Auth().datacube(...)` function now manage multiple collections.
+- Uses `fields` to query only assets asked over Skyfox/EDS (enhance performance).
 
 ## [0.0.1-rc9] - 2023-12-12