+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.1] - 2023-12-15
+
+### Added
+
+- Auth datacube `earthdaily.earthdatastore.Auth().datacube(...)` function now manage multiple collections.
+- Uses `fields` to query only assets asked over Skyfox/EDS (enhance performance).
+
+### Changed
+
+- `asset_mapper` has now all assets from available collections.
+
 ## [0.0.1-rc9] - 2023-12-12
 
 ### Added